Implement new media query for Navbar, reorganize styling for 1440px
Implemented separate media query for displays smaller than 1440px in width so
that the 1440px media query section can be simplified. Since the most drastic
changes take place in the smaller display in active state (when navbar menu is
displayed) it makes sense to put that special styling as non-default. Otherwise
it is necessary to undo a lot of that styling in the 1440px media query.